requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
googlemaps>=4.10.0
langgraph==0.6.7
langchain_openai==0.3.28
pydantic
langchain_community==0.3.27
langchain-experimental
wikipedia
langgraph_supervisor
streamlit
winsdk==1.0.0b10
